requirements.txt
aiogram>=3.4.1
Pillow>=10.2.0